DNN_TorchFM_TTower/requirements.txt

numpy>=1.23,<1.28
pandas>=1.5,<2.1
scikit-learn>=1.2,<1.7
tqdm>=4.66

torch==2.2.1          
deepctr-torch==0.2.9  

psycopg2-binary>=2.9